pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gm-connect"
version = "0.1.0"
description = "A Python package to send, read, and manage Gmail via IMAP/SMTP with CLI support."
readme = "README.md"
authors = [
    { name="Nitesh Mhatre", email="n.mhatre90@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/nitesh-mhatre/gm-connect"
Issues = "https://github.com/nitesh-mhatre/gm-connect/issues"

[tool.setuptools.packages.find]
where = ["src"]